checkbox_detector_gemini/requirements.txt
google-genai>=1.0.0
opencv-python>=4.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
Pillow>=10.0.0
supervision>=0.20.0

